refactor(browser): Replace monolithic browser tools with modular design
- Replace single browser_tools.py file with modular architecture
- Create separate files for service layer, MCP clients, tools, and types
- Update README.md and __init__.py for the new structure
- Improve code organization with clearer separation of concerns
